refactor(guardrails): split GuardedProvider helpers into peer files
internal/guardrails/provider.go was 1569 lines. The GuardedProvider
decorator is small; the bulk was three other concerns (raw JSON
batch-body rewriting, chat envelope preservation, responses input
preservation). Split into peer files in the same package:

  provider.go (512)                  GuardedProvider struct + all
                                     decorator methods, message-clone helpers
  batch_rewrite.go (188)             rewriteGuardedChatBatchBody,
                                     patchGuardedChatBatchBody,
                                     patchChatMessagesJSON, patchRawChatMessage,
                                     rewriteGuardedResponsesBatchBody,
                                     patchJSONObjectFields + jsonFieldPatch,
                                     unmarshalJSONArray, isZeroJSONFieldValue
  chat_message_apply.go (250)        chatToMessages, applyMessagesToChat-
                                     PreservingEnvelope, tailMatchedSystem-
                                     Offsets, applyGuardedMessageToOriginal,
                                     newChatMessageFromGuardrail,
                                     applyGuardedContentToOriginal,
                                     rewriteStructuredContentWithTextRewrite,
                                     normalizeGuardrailMessageText
  responses_message_apply.go (641)   responsesToMessages, responsesInput-
                                     ToMessages, coerceResponsesInputElements,
                                     responsesInputElementToGuardrailMessage,
                                     applyMessagesToResponses + envelope/element
                                     patchers, structured-content rewriters,
                                     cloneStringAnyMap

Pure relocation; tests pass.
